test: 아티스트 검색 로직 분리로 인한 수정
