+++ src/test/java/com/birca/bircabackend/support/enviroment/DocumentationTest.java
@@ -93,6 +93,9 @@ public class DocumentationTest {
     @MockBean
     protected CafeQueryService cafeQueryService;
 
+    @MockBean
+    protected ArtistSearchQueryService artistSearchQueryService;
+
     protected List<FieldDescriptor> getErrorDescriptor(ErrorCode[] errorCodes) {
         return Arrays.stream(errorCodes)
                 .flatMap(errorCode -> Stream.of(
